requirements.txt
Faker
python-dotenv
onesecmail
Pillow
keras-ocr
injector
tensorflow
random-username
MailTm
unidecode
bs4
psutil
